15.10.24-22:37

+++ DataBase/Models/Product.cs
@@ -11,6 +11,6 @@
         public Brand Brand { get; set; }
         public Model Model { get; set; }
         public List<Property> Properties { get; set; }
-        public ushort Price { get; set; }
+        public decimal Price { get; set; }
     }
 }
